Fix Home - Task is not created in Concierge after sending [] task from Concierge prompt box
