+++ src/pages/inbox/report/ReportActionCompose/useComposerSubmit.ts
@@ -7,18 +7,13 @@ import usePermissions from '@hooks/usePermissions';
 import useReportIsArchived from '@hooks/useReportIsArchived';
 
 import {addAttachmentWithComment, addComment, clearAgentZeroProcessingIndicator} from '@libs/actions/Report';
-import {createTaskAndNavigate, setNewOptimisticAssignee} from '@libs/actions/Task';
-import {isEmailPublicDomain} from '@libs/LoginUtils';
+import {createTaskFromMarkdown} from '@libs/actions/Task';
 import {rand64} from '@libs/NumberUtils';
-import {addDomainToShortMention} from '@libs/ParsingUtils';
-import {getPersonalDetailByEmail} from '@libs/PersonalDetailsUtils';
 import {getAllReportActions} from '@libs/ReportActionsUtils';
 import {canUserPerformWriteAction, generateReportID, isConciergeChatReport} from '@libs/ReportUtils';
-import {getAllPersonalDetailLogins} from '@libs/ShortMentionLogins';
 import {startSpan} from '@libs/telemetry/activeSpans';
 import getSendMessageListWeight from '@libs/telemetry/getSendMessageListWeight';
 import getSendMessageSource from '@libs/telemetry/getSendMessageSource';
-import {generateAccountID} from '@libs/UserUtils';
 
 import {useActionListContext} from '@pages/inbox/ActionListContext';
 
@@ -26,12 +21,8 @@ import {setIsComposerFullSize} from '@userActions/Report';
 
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
-import type * as OnyxTypes from '@src/types/onyx';
-
-import type {OnyxEntry} from 'react-native-onyx';
 
 import {useRoute} from '@react-navigation/native';
-import {Str} from 'expensify-common';
 
 import {useComposerActions, useComposerEditActions, useComposerEditState, useComposerMeta, useComposerSendState} from './ComposerContext';
 import useComposerReportData from './useComposerReportData';
@@ -61,8 +52,6 @@ function useComposerSubmit(reportID: string) {
 
     const reportAncestors = useAncestors(report);
     const targetReportAncestors = useAncestors(targetReport);
-
-    const currentUserEmail = currentUserPersonalDetails.email ?? '';
 
     /**
      * Add or edit a comment in the composer
@@ -107,57 +96,8 @@ function useComposerSubmit(reportID: string) {
             return;
         }
 
-        const taskMatch = draftMessageTrimmed.match(CONST.REGEX.TASK_TITLE_WITH_OPTIONAL_SHORT_MENTION);
-        if (taskMatch) {
-            let taskTitle = taskMatch[3] ? taskMatch[3].trim().replaceAll('\n', ' ') : undefined;
-            if (taskTitle) {
-                const mention = taskMatch[1] ? taskMatch[1].trim() : '';
-                const currentUserPrivateDomain = isEmailPublicDomain(currentUserEmail) ? '' : Str.extractEmailDomain(currentUserEmail);
-                const mentionWithDomain = addDomainToShortMention(mention, getAllPersonalDetailLogins(), currentUserPrivateDomain) ?? mention;
-                const isValidMention = Str.isValidEmail(mentionWithDomain);
-
-                let assignee: OnyxEntry<OnyxTypes.PersonalDetails>;
-                let assigneeChatReport;
-                if (mentionWithDomain) {
-                    if (isValidMention) {
-                        assignee = getPersonalDetailByEmail(mentionWithDomain);
-                        if (!assignee) {
-                            const optimisticDataForNewAssignee = setNewOptimisticAssignee(currentUserPersonalDetails.accountID, {
-                                accountID: generateAccountID(mentionWithDomain),
-                                login: mentionWithDomain,
-                            });
-                            assignee = optimisticDataForNewAssignee.assignee;
-                            assigneeChatReport = optimisticDataForNewAssignee.assigneeReport;
-                        }
-                    } else {
-                        taskTitle = `@${mentionWithDomain} ${taskTitle}`;
-                    }
-                }
-
-                const taskCreatorAndAssigneeDetails = {[currentUserPersonalDetails.accountID]: currentUserPersonalDetails};
-                if (assignee) {
-                    taskCreatorAndAssigneeDetails[assignee.accountID] = assignee;
-                }
-
-                createTaskAndNavigate({
-                    parentReport: report,
-                    title: taskTitle,
-                    description: '',
-                    assigneeEmail: assignee?.login ?? '',
-                    currentUserAccountID: currentUserPersonalDetails.accountID,
-                    currentUserEmail,
-                    currentUserDisplayName: currentUserPersonalDetails.displayName,
-                    currentUserAvatar: currentUserPersonalDetails.avatar,
-                    assigneeAccountID: assignee?.accountID,
-                    assigneeChatReport,
-                    policyID: report?.policyID,
-                    isCreatedUsingMarkdown: true,
-                    quickAction,
-                    ancestors: reportAncestors,
-                    taskCreatorAndAssigneeDetails,
-                });
-                return;
-            }
+        if (createTaskFromMarkdown({text: draftMessageTrimmed, parentReport: report, currentUserPersonalDetails, quickAction, ancestors: reportAncestors})) {
+            return;
         }
 
         const optimisticReportActionID = rand64();
